requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.23
PyMySQL==1.0.2
baidu-aip==2.2.18.0
python-dotenv==0.19.0
gunicorn==23.0.0
requests==2.31.0 